requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.10
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dotenv==1.2.1
sqlparse==0.5.4
